requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==4.0.7
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-timezone-field==5.0
djangorestframework==3.14.0
func-timeout==4.3.5
gunicorn==20.1.0
idna==3.4
kombu==5.2.4
numpy==1.23.3
packaging==21.3
pandas==1.5.0
prompt-toolkit==3.0.31
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wget==3.2
wrapt==1.14.1
